requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
idna==3.3
numpy==1.21.4
opencv-python==4.5.4.60
Pillow==8.4.0
requests==2.26.0
urllib3==1.26.7
